requirements.txt
# Defeat Beta API for earnings data
defeatbeta-api

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# LLM Integration
ollama>=0.1.0

# Configuration and utilities
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0

# RAG dependencies
sentence-transformers>=2.2.0
pgvector>=0.2.0